No longer pulling in bower configs separately
This is a breaking change for previous projects that used bsp-grunt, so going up a full version. This will no longer copy bower files for you via the bower config. Now, the bower dependencies main entries are pulled into the root of both the styles and scripts folders. You can pull them in in your respective main.less and main.js from there.